+++ WireAuthentication/Sources/WireAuthentication/Components/DetermineAuthMethodComponent.swift
@@ -73,8 +73,8 @@ class DetermineAuthMethodComponent: Component<DetermineAuthMethodComponentDepend
     func loginViaEmailComponent(
         email: String?,
         backendMetadata: BackendMetadata
-    ) -> LoginViaEmailNewComponent {
-        LoginViaEmailNewComponent(
+    ) -> LoginViaEmailComponent {
+        LoginViaEmailComponent(
             parent: self,
             email: email,
             environmentType: dependency.environmentType,
@@ -164,14 +164,14 @@ extension DetermineAuthMethodComponent: DetermineAuthMethodViewModel.Factory {
 extension DetermineAuthMethodComponent: DetermineAuthMethodView.Factory {
 
     @MainActor
-    func loginViaEmailViewNew(
+    func loginViaEmailView(
         email: String?,
         canCreateAccount: Bool,
         didDetectDomainConflict: Bool,
         environmentType: BackendEnvironmentType,
         backendConfig: BackendConfig,
         backendMetadata: BackendMetadata
-    ) -> LoginViaEmailViewNew {
+    ) -> LoginViaEmailView {
         loginViaEmailComponent(
             email: email,
             backendMetadata: backendMetadata
@@ -180,22 +180,6 @@ extension DetermineAuthMethodComponent: DetermineAuthMethodView.Factory {
             didDetectDomainConflict: false
         )
     }
-
-//    @MainActor
-//    func loginViaEmailView(
-//        email: String,
-//        canCreateAccount: Bool,
-//        didDetectDomainConflict: Bool,
-//        environmentType: BackendEnvironmentType,
-//        backendConfig: BackendConfig,
-//        backendMetadata: BackendMetadata
-//    ) -> LoginViaEmailView {
-//        loginViaEmailComponent(backendMetadata: backendMetadata).view(
-//            email: email,
-//            canCreateAccount: canCreateAccount,
-//            didDetectDomainConflict: didDetectDomainConflict
-//        )
-//    }
 
     func loginViaSSOView(
         ssoURL: URL,
