+++ WireAuthentication/Sources/WireAuthentication/Components/DetermineAuthMethodOnPremComponent.swift
@@ -83,26 +83,13 @@ class DetermineAuthMethodOnPremComponent: Component<DetermineAuthMethodOnPremCom
 
     // MARK: - Children
 
-//    func loginViaEmailComponent(
-//        environmentType: BackendEnvironmentType,
-//        backendConfig: BackendConfig,
-//        backendMetadata: BackendMetadata
-//    ) -> LoginViaEmailComponent {
-//        LoginViaEmailComponent(
-//            parent: self,
-//            environmentType: environmentType,
-//            backendConfig: backendConfig,
-//            backendMetadata: backendMetadata
-//        )
-//    }
-
     func loginViaEmailComponent(
         email: String?,
         environmentType: BackendEnvironmentType,
         backendConfig: BackendConfig,
         backendMetadata: BackendMetadata
-    ) -> LoginViaEmailNewComponent {
-        LoginViaEmailNewComponent(
+    ) -> LoginViaEmailComponent {
+        LoginViaEmailComponent(
             parent: self,
             email: email,
             environmentType: environmentType,
@@ -191,14 +178,14 @@ extension DetermineAuthMethodOnPremComponent: DetermineAuthMethodViewModel.Facto
 
 extension DetermineAuthMethodOnPremComponent: DetermineAuthMethodView.Factory {
     @MainActor
-    func loginViaEmailViewNew(
+    func loginViaEmailView(
         email: String?,
         canCreateAccount: Bool,
         didDetectDomainConflict: Bool,
         environmentType: BackendEnvironmentType,
         backendConfig: BackendConfig,
         backendMetadata: BackendMetadata
-    ) -> LoginViaEmailViewNew {
+    ) -> LoginViaEmailView {
         loginViaEmailComponent(
             email: email,
             environmentType: environmentType,
@@ -210,26 +197,6 @@ extension DetermineAuthMethodOnPremComponent: DetermineAuthMethodView.Factory {
         )
     }
 
-
-//    @MainActor
-//    func loginViaEmailView(
-//        email: String,
-//        canCreateAccount: Bool,
-//        didDetectDomainConflict: Bool,
-//        environmentType: BackendEnvironmentType,
-//        backendConfig: BackendConfig,
-//        backendMetadata: BackendMetadata
-//    ) -> LoginViaEmailView {
-//        loginViaEmailComponent(
-//            environmentType: environmentType,
-//            backendConfig: backendConfig,
-//            backendMetadata: backendMetadata
-//        ).view(
-//            email: email,
-//            canCreateAccount: canCreateAccount,
-//            didDetectDomainConflict: didDetectDomainConflict
-//        )
-//    }
 
     func loginViaSSOView(
         ssoURL: URL,
